requirements.txt
python-telegram-bot==20.3 
SQLAlchemy==1.4.23 
flask==2.0.1 
requests==2.28.1 
python-dotenv==0.19.0 
pyTelegramBotAPI==4.12.0 
nest-asyncio==1.5.6 
PySocks==1.7.1 
werkzeug==2.0.3 
telethon==1.28.5